Code cleanup (#71)
* Added sonarQube code analysis

* Fixed the sonar with the mysql server that was missing

* Added SonarQube code analysis


* Added a default value to the expiryDate if the env var is not found.


* Finish the delete of the User entity and the Folder entity and the file that are contained in it


* Refactored all the code to comply with the SonarLinter


* fixed an issue with a merge conflict


* Fixed an algorithm issue


* Changed a property in application.properties

---------

+++ src/main/java/fr/nil/backedflow/config/WebConfig.java
@@ -25,7 +25,7 @@ public class WebConfig implements WebMvcConfigurer {
                 .allowedOrigins("https://api.transfer-flow.studio")
                 .exposedHeaders("Access-Control-Allow-Origin")
                 .allowCredentials(true)
-                .maxAge(3600*5)
+                .maxAge(((long) 3600 * 5))
                 .allowedMethods("GET", "POST", "PUT", "DELETE", "HEAD");
 
     }
